requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
cytoolz==0.9.0
dash==0.19.0
dash-core-components==0.16.0rc1
dash-html-components==0.9.0rc1
dash-renderer==0.13.0rc2
decorator==4.1.2
dill==0.2.5
exps==0.1
Flask==0.12.2
Flask-Compress==1.4.0
fn==0.4.3
future==0.16.0
idna==2.6
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.0
nbformat==4.4.0
numpy==1.14.0rc1
pandas==0.21.1
plotly==2.2.3
PyFunctional==1.0.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
six==1.10.0
SQLAlchemy==1.2.0
tabulate==0.7.7
toolz==0.9.0
traitlets==4.3.2
urllib3==1.22
voluptuous==0.10.5
Werkzeug==0.13
